Remove boilerplate for attribute construction in loaders
- programs.js: extend context with initAttributes method, for
  constructing multiple attributes based on an attrInfo object
- */program.js: use new initAttributes method
- fill/program.js: incorporate loader.js code
